Split Params into ui-params
