+++ packages/app-democracy/src/Referendum.tsx
@@ -3,14 +3,14 @@
 // of the Apache-2.0 license. See the LICENSE file for details.
 
 import { I18nProps } from '@polkadot/ui-app/types';
-import { RawParam } from '@polkadot/ui-app/Params/types';
+import { RawParam } from '@polkadot/ui-params/types';
 import { DerivedReferendumVote } from '@polkadot/ui-react-rx/derive/types';
 
 import BN from 'bn.js';
 import React from 'react';
 import { ReferendumInfo } from '@polkadot/types';
 import { Chart, Static } from '@polkadot/ui-app/index';
-import VoteThreshold from '@polkadot/ui-app/Params/Param/VoteThreshold';
+import VoteThreshold from '@polkadot/ui-params/Param/VoteThreshold';
 import { withCall, withMulti } from '@polkadot/ui-react-rx/with/index';
 import { balanceFormat, numberFormat } from '@polkadot/ui-react-rx/util/index';
 import settings from '@polkadot/ui-settings';
