+++ packages/app-explorer/src/BlockByHash/Logs.tsx
@@ -7,7 +7,7 @@ import { I18nProps } from '@polkadot/ui-app/types';
 import React from 'react';
 import { Struct, Tuple, Vector, getTypeDef } from '@polkadot/types/codec';
 import { DigestItem } from '@polkadot/types/Digest';
-import { Params } from '@polkadot/ui-app/index';
+import Params from '@polkadot/ui-params/index';
 
 import translate from '../translate';
