+++ packages/app-extrinsics/src/Params/Call.tsx
@@ -2,7 +2,7 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import { Props as BaseProps } from '@polkadot/ui-app/Params/types';
+import { Props as BaseProps } from '@polkadot/ui-params/types';
 import { ApiProps } from '@polkadot/ui-react-rx/types';
 
 import React from 'react';
