+++ packages/app-storage/src/Query.tsx
@@ -9,7 +9,7 @@ import { QueryTypes, StorageModuleQuery } from './types';
 import React from 'react';
 import { Compact } from '@polkadot/types/codec';
 import { Button, Labelled } from '@polkadot/ui-app/index';
-import valueToText from '@polkadot/ui-app/Params/valueToText';
+import valueToText from '@polkadot/ui-params/valueToText';
 import { withCallDiv } from '@polkadot/ui-react-rx/with/index';
 import { isU8a, u8aToHex, u8aToString } from '@polkadot/util';
