+++ packages/ui-params/README.md
@@ -1,0 +1,1 @@
+# @polkadot/ui-params
