+++ packages/ui-params/src/Param/Amount.tsx
@@ -4,55 +4,54 @@
 
 import { Props } from '../types';
 
+import BN from 'bn.js';
 import React from 'react';
-import { decodeAddress } from '@polkadot/keyring';
+import { Input } from '@polkadot/ui-app/index';
+import numberFormat from '@polkadot/ui-react-rx/util/numberFormat';
 
-import InputAddress from '../../InputAddress';
 import Bare from './Bare';
 
-export default class Account extends React.PureComponent<Props> {
+export default class Amount extends React.PureComponent<Props> {
   render () {
     const { className, defaultValue: { value }, isDisabled, isError, label, style, withLabel } = this.props;
-    const defaultValue = value && value.toString();
+    const defaultValue = isDisabled
+      ? numberFormat(value)
+      : (
+        value instanceof BN
+          ? value.toNumber()
+          : new BN((value as number) || 0).toNumber()
+      );
 
     return (
       <Bare
         className={className}
         style={style}
       >
-        <InputAddress
-          className={isDisabled ? 'full' : 'large'}
+        <Input
+          className={isDisabled ? 'full' : 'small'}
           defaultValue={defaultValue}
           isDisabled={isDisabled}
           isError={isError}
-          isInput
           label={label}
+          min={0}
           onChange={this.onChange}
-          placeholder='5...'
+          type={
+            isDisabled
+              ? 'text'
+              : 'number'
+          }
           withLabel={withLabel}
         />
       </Bare>
     );
   }
 
-  private onChange = (value?: string): void => {
+  private onChange = (value: string): void => {
     const { onChange } = this.props;
 
-    let isValid = false;
-
-    if (value) {
-      try {
-        decodeAddress(value);
-
-        isValid = true;
-      } catch (err) {
-        console.error(err);
-      }
-    }
-
     onChange && onChange({
-      isValid,
-      value
+      isValid: true,
+      value: new BN(value || 0)
     });
   }
 }
