+++ packages/ui-params/src/Param/Bare.tsx
@@ -2,11 +2,10 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import { BareProps } from '../../types';
+import { BareProps } from '@polkadot/ui-app/types';
 
 import React from 'react';
-
-import classes from '../../util/classes';
+import classes from '@polkadot/ui-app/util/classes';
 
 type Props = BareProps & {
   children: React.ReactNode
