+++ packages/ui-params/src/Param/File.tsx
@@ -2,11 +2,11 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import { BareProps } from '../../types';
+import { BareProps } from '@polkadot/ui-app/types';
 
 import React from 'react';
+import { InputFile } from '@polkadot/ui-app/index';
 
-import InputFile from '../../InputFile';
 import Bare from './Bare';
 
 type Props = BareProps & {
