+++ packages/ui-params/src/Param/Proposal.tsx
@@ -6,10 +6,9 @@ import { Props } from '../types';
 
 import React from 'react';
 import { Extrinsic, Method } from '@polkadot/types';
+import { Call, Static } from '@polkadot/ui-app/index';
+import classes from '@polkadot/ui-app/util/classes';
 
-import Call from '../../Call';
-import Static from '../../Static';
-import classes from '../../util/classes';
 import Bare from './Bare';
 import Unknown from './Unknown';
