+++ packages/ui-params/src/Param/index.tsx
@@ -2,14 +2,14 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import { I18nProps } from '../../types';
+import { I18nProps } from '@polkadot/ui-app/types';
 import { BaseProps, Props as ComponentProps, ComponentMap } from '../types';
 
 import React from 'react';
+import classes from '@polkadot/ui-app/util/classes';
+import translate from '@polkadot/ui-app/translate';
 import { isUndefined } from '@polkadot/util';
 
-import classes from '../../util/classes';
-import translate from '../../translate';
 import findComponent from './findComponent';
 
 type Props = I18nProps & BaseProps & {
